Rotating now updates the highlighted tile
All of the other tiles flash the current connectors for a fraction of a second but it works for now. Also added a buffer to when the tile stops displaying the current selection so that it doesn't flash while hovering over the same tile. Could use some tuning and maybe a different approach later, but again it works for now.